Split out metric storing and querying into separate source files
